Split ProgramStateTest
